[rl] Accumulate Grug expert outputs in FP32
Accumulate weighted BF16 expert outputs in FP32 before casting the MoE result back to model dtype. This removes fixed-route eager/grouped reduction-order drift while leaving routing and expert computation unchanged.

A route-aware 32-H100 replay passed 12,287 of 12,288 sampled action log probabilities; the one miss exceeded its allowance by 0.00070. Global CE and 5,184 sampled gradients passed. An eight-H100 fixed-route measurement found 3.84% slower block forward/backward and 251,592,704 additional bytes at the combine boundary.

+++ skyrl-train/skyrl_train/models/grug_moe.py
@@ -363,7 +363,7 @@ class GrugMoeExperts(nn.Module):
         selected_experts: torch.Tensor,
         combine_weights: torch.Tensor,
     ) -> torch.Tensor:
-        output = torch.zeros_like(hidden_states)
+        output = torch.zeros_like(hidden_states, dtype=torch.float32)
         for expert_idx in range(self.num_experts):
             token_idx, slot_idx = torch.where(selected_experts == expert_idx)
             if token_idx.numel() == 0:
@@ -373,8 +373,8 @@ class GrugMoeExperts(nn.Module):
             up = F.linear(expert_input, self.up_proj.weight[expert_idx])
             expert_output = F.linear(F.silu(gate) * up, self.down_proj.weight[expert_idx])
             weight = combine_weights[token_idx, slot_idx].unsqueeze(-1).to(expert_output.dtype)
-            output.index_add_(0, token_idx, expert_output * weight)
-        return output
+            output.index_add_(0, token_idx, (expert_output * weight).float())
+        return output.to(hidden_states.dtype)
 
     def forward(
         self,
@@ -509,14 +509,18 @@ class _GrugGroupedExpertExecution:
         routed_indices, routed_output = grouped_expert_contributions(
             experts,
             hidden_states,
-            combine_weights,
+            combine_weights.to(hidden_states.dtype),
             selected_experts,
             reorderer,
         )
-        return torch.zeros_like(hidden_states).scatter_add(
-            dim=0,
-            index=routed_indices,
-            src=routed_output,
+        return (
+            torch.zeros_like(hidden_states, dtype=torch.float32)
+            .scatter_add(
+                dim=0,
+                index=routed_indices,
+                src=routed_output.float(),
+            )
+            .to(hidden_states.dtype)
         )
